[step1-2] 지하철 노선 생성 기능 구현

+++ src/main/java/subway/line/CreateLineRequest.java
@@ -3,19 +3,19 @@ package subway.line;
 public class CreateLineRequest {
     private final String name;
     private final String color;
-    private final int upStationId;
-    private final int downStationId;
+    private final Long upStationId;
+    private final Long downStationId;
     private final int distance;
 
     public CreateLineRequest() {
         this.name = null;
         this.color = null;
-        this.upStationId = 0;
-        this.downStationId = 0;
+        this.upStationId = 0L;
+        this.downStationId = 0L;
         this.distance = 0;
     }
 
-    public CreateLineRequest(String name, String color, int upStationId, int downStationId, int distance) {
+    public CreateLineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
@@ -31,11 +31,11 @@ public class CreateLineRequest {
         return color;
     }
 
-    public int getUpStationId() {
+    public Long getUpStationId() {
         return upStationId;
     }
 
-    public int getDownStationId() {
+    public Long getDownStationId() {
         return downStationId;
     }
